# refactor(datascience): use functions to evaluate model performance