--- a/datascience/scripts/train_image_model.py
+++ b/datascience/scripts/train_image_model.py
@@ -1,9 +1,6 @@
 # type: ignore
 """Train image model, evaluate its performance and generates figures and stats.
 
-from shutil import ignore_patterns
-
-
 All logs and best checkpoints are stored in --output-dir.
 --input-dir expects a dataset with this structure is expected:
 
@@ -12,42 +9,50 @@ dataset_dir
 ├── X_train.csv
 ├── images
 │   ├── test
-│   │   └── 2705 # One folder per category id
-│   │       └── image_977803476_product_278535420.jpg
+│   │   └── image_977803476_product_278535420.jpg
 │   └── train
-│       └── 2583 # One folder per category id
-│           └── image_1174586892_product_2940638801.jpg
+│       └── image_1174586892_product_2940638801.jpg
 ├── y_test.csv
 └── y_train.csv
 
 Use scripts/optimize_images.py to generate a dataset.
 """
-import itertools
 import logging
 import os
 import sys
 from pathlib import Path
 
-import matplotlib.pyplot as plt
+import keras
 import numpy as np
 import pandas as pd
 import pydantic
 import pydantic_argparse
 import tensorflow as tf
-from pydantic import BaseModel, DirectoryPath
-from sklearn import metrics
-from tensorflow.keras import Sequential, layers  # pyright: ignore
-from tensorflow.keras.callbacks import (  # pyright: ignore
+from keras import Sequential, layers
+from keras.callbacks import (
     CSVLogger,
     EarlyStopping,
     ModelCheckpoint,
     TensorBoard,
 )
-from tensorflow.keras.losses import CategoricalCrossentropy  # pyright: ignore
-from tensorflow.keras.optimizers import SGD  # pyright: ignore
-from tensorflow.train import latest_checkpoint  # pyright: ignore
+from keras.losses import CategoricalCrossentropy
+from keras.optimizers import SGD
+from pydantic import BaseModel, DirectoryPath
+from sklearn import metrics
+from tensorflow.train import latest_checkpoint
 
 from src.core.settings import get_mobilenet_image_model_settings
+from src.utilities.dataset_utils import (
+    get_imgs_filenames,
+    to_img_feature_target,
+    to_normal_category_id,
+    to_simplified_category_id,
+)
+from src.utilities.model_eval import (
+    gen_classification_report,
+    gen_confusion_matrix,
+    gen_training_history_figure,
+)
 
 
 class TrainImageModelArgs(BaseModel):
@@ -107,36 +112,63 @@ def main(args: TrainImageModelArgs) -> int:  # noqa: PLR0915
     # Especially annoying with tensorflow-metal
     os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
 
-    model_settings = get_mobilenet_image_model_settings()
+    logger.info("Load data")
+
+    df_X_train = pd.read_csv(args.input_dir / "X_train.csv", index_col=0).fillna("")
+    df_X_test = pd.read_csv(args.input_dir / "X_test.csv", index_col=0).fillna("")
+
+    logger.info("Extract features and target")
+    X_train = get_imgs_filenames(
+        df_X_train["productid"].to_list(),
+        df_X_train["imageid"].to_list(),
+        args.input_dir / "images",
+    )
+    X_test = get_imgs_filenames(
+        df_X_test["productid"].to_list(),
+        df_X_test["imageid"].to_list(),
+        args.input_dir / "images",
+    )
+
+    y_train = np.loadtxt(
+        args.input_dir / "y_train.csv",
+        dtype=int,
+        delimiter=",",
+        skiprows=1,
+        usecols=(1),
+    )
+    y_test = np.loadtxt(
+        args.input_dir / "y_test.csv", dtype=int, delimiter=",", skiprows=1, usecols=(1)
+    )
+
+    y_train_simplified = to_simplified_category_id(y_train)
+    y_test_simplified = to_simplified_category_id(y_test)
+
+    y_train_categorical = keras.utils.to_categorical(y_train_simplified)
+    y_test_categorical = keras.utils.to_categorical(y_test_simplified)
 
     logger.info("Create datasets")
-
-    train_dir = args.input_dir / "images" / "train"
-    test_dir = args.input_dir / "images" / "test"
-    (train_ds, val_ds) = tf.keras.utils.image_dataset_from_directory(
-        train_dir,
-        validation_split=0.2,
-        subset="both",
-        seed=args.seed,
-        image_size=(model_settings.IMG_WIDTH, model_settings.IMG_HEIGHT),
-        batch_size=args.batch_size,
-        label_mode="categorical",
+    train_ds = (
+        tf.data.Dataset.from_tensor_slices((X_train, y_train_categorical))
+        .map(to_img_feature_target)
+        .batch(args.batch_size)
     )
 
-    nb_train_images = len(train_ds.file_paths)
-    nb_val_images = len(val_ds.file_paths)
-
-    test_ds = tf.keras.utils.image_dataset_from_directory(
-        test_dir,
-        image_size=(model_settings.IMG_WIDTH, model_settings.IMG_HEIGHT),
-        batch_size=args.batch_size,
-        label_mode="categorical",
+    test_ds = (
+        tf.data.Dataset.from_tensor_slices((X_test, y_test_categorical))
+        .map(to_img_feature_target)
+        .batch(args.batch_size)
     )
+
+    model_settings = get_mobilenet_image_model_settings()
+
+    nb_train_images = len(X_train)
+    nb_test_images = len(X_test)
+
     # Add cache configuration to speed up training
     # If this cause issues, we'll add an argument to enable it
     autotune = tf.data.AUTOTUNE
     train_ds = train_ds.cache().prefetch(buffer_size=autotune)
-    val_ds = val_ds.cache().prefetch(buffer_size=autotune)
+    test_ds = test_ds.cache().prefetch(buffer_size=autotune)
 
     logger.info("Build image model")
     mobilenet_layers = tf.keras.applications.MobileNetV2(
@@ -221,19 +253,19 @@ def main(args: TrainImageModelArgs) -> int:  # noqa: PLR0915
         # Insert the metrics into a CSV file
         CSVLogger(history_file_path, separator=",", append=True),
         # Log information to display them in TensorBoard
-        TensorBoard(log_dir=tensorboard_logs_dir, histogram_freq=1),
+        TensorBoard(log_dir=str(tensorboard_logs_dir), histogram_freq=1),
     ]
 
     logger.info("Start model training")
     steps_per_epoch = nb_train_images // args.batch_size
-    validation_steps = nb_val_images // args.batch_size
+    validation_steps = nb_test_images // args.batch_size
 
     # Train the top layer
     model.fit(
         train_ds.repeat(),
         epochs=100,
         steps_per_epoch=steps_per_epoch,
-        validation_data=val_ds.repeat(),
+        validation_data=test_ds.repeat(),
         validation_steps=validation_steps,
         callbacks=cp_callbacks,
     )
@@ -247,7 +279,7 @@ def main(args: TrainImageModelArgs) -> int:  # noqa: PLR0915
     mobilenet_layers.trainable = True
     model.compile(
         optimizer=SGD(learning_rate=0.001, momentum=0.9),
-        loss=CategoricalCrossentropy(from_logits=True, label_smoothing=0.1),
+        loss=CategoricalCrossentropy(label_smoothing=0.1),
         metrics=["accuracy"],
     )
     model.summary()
@@ -255,7 +287,7 @@ def main(args: TrainImageModelArgs) -> int:  # noqa: PLR0915
         train_ds.repeat(),
         epochs=100,
         steps_per_epoch=steps_per_epoch,
-        validation_data=val_ds.repeat(),
+        validation_data=test_ds.repeat(),
         validation_steps=validation_steps,
         callbacks=cp_callbacks,
     )
@@ -267,124 +299,24 @@ def main(args: TrainImageModelArgs) -> int:  # noqa: PLR0915
         model.load_weights(latest)
     model.save(args.output_dir / "cnn_mobilenetv2.keras")
 
-    # TODO @joffreylgt: figure and stats generation should be in their own functions.
-    #  https://github.com/JoffreyLGT/e-commerce-mlops/issues/103
-
     logger.info("Generate training history figure")
-    training_history = pd.read_csv(history_file_path, delimiter=",", header=0)
+    logger.info(gen_training_history_figure(history_file_path, args.output_dir))
 
-    fig = plt.figure(figsize=(10, 3))
-    ax1 = fig.add_subplot(121)
+    logger.info("Predict test data categories")
+    y_pred_simplified = model.predict(test_ds)
+    y_pred = to_normal_category_id([np.argmax(i) for i in y_pred_simplified])
 
-    # TODO @joffreylgt: translate from French to English.
-    #  https://github.com/JoffreyLGT/e-commerce-mlops/issues/103
+    logger.info(f"Accuracy score: {metrics.accuracy_score(y_test, y_pred)}")
 
-    # Labels des axes
-    ax1.set_xlabel("Epochs")
-    ax1.set_ylabel("Accuracy")
-
-    # Courbe de la précision sur l'échantillon d'entrainement
-    ax1.plot(
-        np.arange(1, training_history["accuracy"].count() + 1, 1),
-        training_history["accuracy"],
-        label="Training Accuracy",
-        color="blue",
-    )
-
-    # Courbe de la précision sur l'échantillon de test
-    ax1.plot(
-        np.arange(1, training_history["val_accuracy"].count() + 1, 1),
-        training_history["val_accuracy"],
-        label="Validation Accuracy",
-        color="red",
-    )
-
-    ax1.legend()
-    ax1.set_title("Accuracy per epoch")
-
-    ax2 = fig.add_subplot(122)
-    ax2.set_ylabel("Loss")
-    ax2.set_xlabel("Epochs")
-
-    ax2.plot(
-        np.arange(1, training_history["loss"].count() + 1, 1),
-        training_history["loss"],
-        label="Training loss",
-        linestyle="dashed",
-        color="blue",
-    )
-
-    ax2.plot(
-        np.arange(1, training_history["val_loss"].count() + 1, 1),
-        training_history["val_loss"],
-        label="Validation loss",
-        linestyle="dashed",
-        color="red",
-    )
-
-    ax2.legend()
-    ax2.set_title("Loss per epoch")
-
-    plt.savefig(args.output_dir / "training_history.png")
-
-    logger.info("Predict category on test data")
-
-    predictions = np.array([])
-    labels = np.array([])
-    for x, y in test_ds:
-        predictions = np.concatenate(
-            [predictions, np.argmax(model.predict(x, verbose=0), axis=-1)]
-        )
-        labels = np.concatenate([labels, np.argmax(y.numpy(), axis=-1)])
-
-    prdtypecodes = sorted([int(x) for x in os.listdir(test_dir)])
-
-    test_labels = [prdtypecodes[int(i)] for i in labels]
-    predictions_labels = [prdtypecodes[int(i)] for i in predictions]
-
-    logger.info(
-        f"Accuracy score: {metrics.accuracy_score(test_labels, predictions_labels)}"
-    )
     logger.info("Generate classification report")
-    class_report = metrics.classification_report(
-        test_labels, predictions_labels, zero_division=0.0  # pyright: ignore
-    )
-
-    Path(args.output_dir / "classification_report.txt").write_text(str(class_report))
-    logger.info(str(class_report))
+    (_, class_report) = gen_classification_report(y_test, y_pred, args.output_dir)
+    logger.info(class_report)
 
     logger.info("Generate confusion matrix")
-    cnf_matrix = np.round(
-        metrics.confusion_matrix(test_labels, predictions_labels, normalize="true"), 2
-    )
-
-    classes = range(0, nb_output_classes)
-
-    plt.figure(figsize=(13, 13))
-
-    plt.imshow(cnf_matrix, interpolation="nearest", cmap="Blues")
-    plt.title("Matrice de confusion")
-    tick_marks = classes
-    plt.xticks(tick_marks, prdtypecodes)
-    plt.yticks(tick_marks, prdtypecodes)
-
-    for i, j in itertools.product(
-        range(cnf_matrix.shape[0]), range(cnf_matrix.shape[1])
-    ):
-        plt.text(
-            j,
-            i,
-            cnf_matrix[i, j],
-            horizontalalignment="center",
-            color="white" if cnf_matrix[i, j] > (cnf_matrix.max() / 2) else "black",
-        )
-
-    plt.ylabel("Vrais labels")
-    plt.xlabel("Labels prédits")
-    plt.xticks(rotation=45)
-    plt.savefig(args.output_dir / "confusion_matrix.png")
+    logger.info(gen_confusion_matrix(y_test, y_pred, args.output_dir))
 
     logger.info("Script finished")
+
     return 0