Refs: #103

# Proof of testing

```txt
❯ python -m scripts.train_text_model --train-patience 2 --input-dir "data/datasets/fusion_model" --output-dir "data/models/text" &> text_training_log.txt

Script started
Load data
Extract features and target
Create preprocessor and preprocess features
Create datasets
Build text model
WARNING in optimizer.__init__, line 70: At this time, the v2.11+ optimizer `tf.keras.optimizers.Adam` runs slowly on M1/M2 Macs, please use the legacy Keras optimizer instead, located at `tf.keras.optimizers.legacy.Adam`.
WARNING in __init__.get, line 292: There is a known slowdown when using v2.11+ Keras optimizers on M1/M2 Macs. Falling back to the legacy Keras optimizer, i.e., `tf.keras.optimizers.legacy.Adam`.
Model: "sequential"
_________________________________________________________________
 Layer (type)                Output Shape              Param #   
=================================================================
 dense (Dense)               (None, 128)               1007232   
                                                                 
 dropout (Dropout)           (None, 128)               0         
                                                                 
 dense_1 (Dense)             (None, 27)                3483      
                                                                 
=================================================================
Total params: 1010715 (3.86 MB)
Trainable params: 1010715 (3.86 MB)
Non-trainable params: 0 (0.00 Byte)
_________________________________________________________________
Load checkpoint: data/models/text/checkpoints/cp_loss-2.06_acc-0.59.ckpt
Start model training
Epoch 1/100

Epoch 1: val_accuracy improved from -inf to 0.58235, saving model to data/models/text/checkpoints/cp_loss-2.05_acc-0.58.ckpt
9/9 - 1s - loss: 0.7142 - accuracy: 0.9976 - val_loss: 2.0540 - val_accuracy: 0.5824 - 810ms/epoch - 90ms/step

Epoch 2/100

Epoch 2: val_accuracy did not improve from 0.58235
9/9 - 0s - loss: 0.6944 - accuracy: 0.9988 - val_loss: 2.0502 - val_accuracy: 0.5800 - 233ms/epoch - 26ms/step

Epoch 3/100

Epoch 3: val_accuracy did not improve from 0.58235
9/9 - 0s - loss: 0.6866 - accuracy: 0.9988 - val_loss: 2.0483 - val_accuracy: 0.5812 - 226ms/epoch - 25ms/step

Save the model
Generate training history figure
data/models/text/training_history.png
Predict test data categories
9/9 - 0s - 209ms/epoch - 23ms/step

Accuracy score: 0.5823529411764706
Generate classification report
              precision    recall  f1-score   support

          10       0.36      0.13      0.19        31
          40       0.48      0.52      0.50        25
          50       0.67      0.71      0.69        17
          60       0.80      1.00      0.89         8
        1140       0.41      0.48      0.44        27
        1160       0.65      0.82      0.73        39
        1180       0.00      0.00      0.00         8
        1280       0.35      0.43      0.39        49
        1281       0.38      0.24      0.29        21
        1300       0.75      0.72      0.73        50
        1301       0.67      0.25      0.36         8
        1302       0.62      0.40      0.49        25
        1320       0.68      0.47      0.56        32
        1560       0.55      0.61      0.58        51
        1920       0.80      0.81      0.80        43
        1940       0.00      0.00      0.00         8
        2060       0.41      0.74      0.53        50
        2220       0.50      0.25      0.33         8
        2280       0.62      0.48      0.54        48
        2403       0.47      0.56      0.51        48
        2462       0.50      0.21      0.30        14
        2522       0.74      0.78      0.76        50
        2582       0.44      0.31      0.36        26
        2583       0.72      0.93      0.81       102
        2585       0.11      0.04      0.06        25
        2705       0.61      0.61      0.61        28
        2905       1.00      0.67      0.80         9

    accuracy                           0.58       850
   macro avg       0.53      0.49      0.49       850
weighted avg       0.56      0.58      0.56       850

Generate confusion matrix
data/models/text/confusion_matrix.png
Script finished


```